chore(integration-tests): Go through axum for engine-v2 integration-tests
Allows us to also test `engine_v2_axum` and tower layers
